updated telescope position and focal length
the telescope positions were given w.r.t. magnetic North, now they are reported w.r.t. geographical North
the effective focal length is now reported, to make it consistent with how it is done in MARS,
the 3cm of change of focal length due to focusing to 10 km is neglected (the effect is tiny either way),
but the factor of 1.0713 due to coma abberation is added